fix: mover format-YAxistTicks.ts de services para utils
